requirements.txt
discord==2.0.0
python-dotenv==0.21.0
xlrd==2.0.1
discord-class-commands==1.1.0